* Changes for vertex (edge is now cost)
