+++ src/algorithms/models/edgecost.ts
@@ -1,0 +1,8 @@
+/**
+ * Edges represent a relation between two vertices.
+ */
+class EdgeCost {
+  distance = 0;
+}
+
+export default EdgeCost;
